MarsRoverV2/requirements.txt
flask==3.0.0
picamera2==0.3.17
Pillow==10.1.0
